refactor: took out logic in HomeForm
