+++ src/components/HomeForm/index.ts
@@ -1,0 +1,1 @@
+export { HomeForm } from "./HomeForm";
